requirements.txt
# Core backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
qrcode==7.4.2
cryptography==41.0.7

# ML & Embeddings (free OSS)
sentence-transformers==2.2.2
faiss-cpu==1.7.4
torch==2.1.0
transformers==4.35.2
numpy==1.24.3
scikit-learn==1.3.2

# Graph & Analytics
networkx==3.2.1
pandas==2.1.3

# HTTP & Integrations
httpx==0.25.2
aiohttp==3.9.1
tenacity==8.2.3

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0
redis==5.0.1
